% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ave_err_commission_and_ommission}
\alias{ave_err_commission_and_ommission}
\title{computes the average number of non-sibs in inferred sibships}
\usage{
ave_err_commission_and_ommission(S, sg.list)
}
\arguments{
\item{S}{A list of inferred sibships where individuals are named by their base-0 indices.}

\item{sg.list}{A list holding the true siblings of each individual.  If
if \code{Ind} is the base-0 index of an individual then
\code{sg.list[[Ind+1]]} is a vector of the base-0 indices of all of its full siblings}
}
\description{
And also computes the average number of sibs missing.
For each individual i in an inferred sibship this function counts the
average number of "errors of commission" (number of individuals
incorrectly declared to be i's sibling.  And it averages over i.
It also computes for each individual i the number of his true sibs
missing from the sibroup he is inferred to be in.
This function operates on a list of sibships
}

