% Generated by roxygen2 (4.0.0): do not edit by hand
\name{count_genos}
\alias{count_genos}
\title{counts up genotypes from a snp_indicators array}
\usage{
count_genos(S)
}
\arguments{
\item{x}{a snp indicators array (3 x L x N)}
}
\description{
counts up genotypes from a snp_indicators array
}
\examples{
snp_genos <- get_snp_genos(fs_dev_test_data$plain_snp_data)
snp_indics <- genos_to_indicators(g = snp_genos$mat)
geno_counts <- count_genos(snp_indics)
}

