% Generated by roxygen2 (4.0.1): do not edit by hand
\name{find_high_logl_sib_pairs}
\alias{find_high_logl_sib_pairs}
\title{for each individual, find the other individuals with a sib-pair log likelihood exceeding a certain amount}
\usage{
find_high_logl_sib_pairs(FSP, UPF, GI, loglV)
}
\arguments{
\item{FSP}{full-sibling pair geno freqs as returned by full_sibling_pair_gfreqs()}

\item{UPF}{unrelated pair geno freqs as would be returned by unrelated_pair_gfreqs()}

\item{GI}{genotypes of all the individuals as a snp_indics array}

\item{loglV}{the log-likelihood value that must be exceeded to be included}
}
\description{
DEPRECATED.  I keep this around to remind me how slow R is.  It is better
to use high_logl_pairs(), implemented in Rcpp, in practice.
}

