\name{full_sib_mcmc}
\alias{full_sib_mcmc}
\title{do a whole MCMC analysis starting from a data set and genotyping error rates}
\usage{
full_sib_mcmc(genos, mu)
}
\arguments{
  \item{genos}{The data frame of SNP genotypes}

  \item{mu}{Genotyping error rates per locus (recycles as
  necessary)}
}
\description{
do a whole MCMC analysis starting from a data set and genotyping error rates
}

