% Generated by roxygen2 (4.0.1): do not edit by hand
\name{geno_post_c}
\alias{geno_post_c}
\title{compute the posterior of Inds genotype given the genotype freqs in the pop}
\usage{
geno_post_c(Gfreqs, Liks)
}
\arguments{
\item{Gfreqs}{array of genotype frequencies}

\item{Liks}{array of likleihoods (3 * L in length)}
}
\description{
compute the posterior of Inds genotype given the genotype freqs in the pop
}

