% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get_marriage_likelihoods}
\alias{get_marriage_likelihoods}
\title{Compute the likelihood of the genotypes of a marriage given the genotype of one of its offspring}
\usage{
get_marriage_likelihoods(offspring_likelihoods, transmission_probs)
}
\arguments{
  \item{offspring_likelihoods}{This is a 3 x L x N array of
  likelihoods of the individual's true genotypes given
  their observed genotypes.  This sort of array is returned
  by the function \code{\link{get_indiv_geno_lik}}.}

  \item{transmission_probs}{a 3 x 3 x 3 matrix of probs of
  kid genotypes given parent genotypes. Such an array is
  returned by \code{\link{trans_probs}}.  See the
  documentation thereof for a description.}
}
\description{
It is advantageous to precompute and store some quantities for pedigree analysis with SNPs.
One of those quantities is the likelihood of a two parents' genotypes given the genotype
of one of the offspring.  This is convenient because the likelihood of the two parents'
genotypes given \emph{all} of their offspring will be the product of the single-offspring
likelihoods.  When genotyping error is present, that has to be thrown into the mix as well.
However, here we assume that the individual offspring genotype likelihoods have already been
computed given their observed genotypes.
}

