% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gibbs_update_one_indiv_in_place}
\alias{gibbs_update_one_indiv_in_place}
\title{must fill in later}
\usage{
gibbs_update_one_indiv_in_place(FSL, IFS, LMMI, LMMFS, PMMFS, KidProngs, Pile,
  Gfreqs, UPG, TP, IndLiks, Ind)
}
\description{
must fill in later
}

