% Generated by roxygen2 (4.0.1): do not edit by hand
\name{greedy_partition}
\alias{greedy_partition}
\title{Uses the estimated posteriors for each sibgroup to come up with a partition into sibgroups}
\usage{
greedy_partition(VS)
}
\arguments{
\item{VS}{A vector of counts of the number of times each sibship was visited.  The sibships are named
according to who is in them via the hashable_sibships function.}
}
\description{
Used internally.  This is a quick and dirty greedy thing. It basically orders sibships
by posterior and then by size, and then it goes through and calls them sibships starting with
the highest posterior ones.  If it reaches one that includes someone that is already in a sibship
it discards it and continues.  At the end, any one not assigned to a sibship is made a singleton.
}

