% Generated by roxygen2 (4.0.0): do not edit by hand
\name{hashable_sibships}
\alias{hashable_sibships}
\title{Returns a vector of strings naming the current sibships}
\usage{
hashable_sibships(S)
}
\arguments{
\item{S}{the full sibling list}
}
\description{
Uses the full sibling list to make names of the sibships like
10-34-67-98, which can be used to hash them
}

