% Generated by roxygen2 (4.0.0): do not edit by hand
\name{high_logl_pairs}
\alias{high_logl_pairs}
\title{return the high-sib-pair-logl indivs for each indiv}
\usage{
high_logl_pairs(FSP, UPF, G, loglV)
}
\arguments{
\item{FSP}{a 9 x L matrix (3 x 3 x L) of expected genotype frequencies of a full sibling pair,
but we pass it in is a vector.}

\item{UPF}{a 9 x L matrix of expected genotype frequencies of an unrelated pair, which we also
pass in as a vector.}

\item{G}{an L x N matrix of 0, 1, or 2 or NA giving the genotypes of the individuals}

\item{loglV}{the cutoff point above which you will accept the pairs.}
}
\description{
I wrote this to see how much faster I could implement the functionality of find_high_logl_sib_pairs
using Rcpp.  find_high_logl_sib_pairs is painfully slow, and I think this should be much faster
}

