% Generated by roxygen2 (4.0.0): do not edit by hand
\name{kid_prongs_times_ind_likelihoods}
\alias{kid_prongs_times_ind_likelihoods}
\title{Computes the posterior predictive likelihoods that an individual with likelihood vector IndGenoLik belongs to each sibling group in AFS}
\usage{
kid_prongs_times_ind_likelihoods(FSL, IndGenoLik, KidProngs, AFS)
}
\arguments{
\item{IndGenoLik}{Likelihoods of the focal individual.  There will be 3*L elements in that.  Note that \code{AFS} is the acceptable
sibling list for this focal individual.}
}
\description{
Note that we have to pass in the whole full sibling list (FSL) so that we can get the LMMI_Idx for each sibling group.
}

