% Generated by roxygen2 (4.0.0): do not edit by hand
\name{lik_array_from_simple_geno_err}
\alias{lik_array_from_simple_geno_err}
\title{makes L 3x3 tables of genotype likelihoods}
\usage{
lik_array_from_simple_geno_err(L, mu)
}
\arguments{
\item{L}{the desired number of loci}

\item{mu}{vector of per-gene-copy genotyping error rates, one per locus.  Will recycle (or be truncated) to be of length L.}
}
\description{
makes L 3x3 tables of genotype likelihoods
}

