% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ordered_fsl}
\alias{ordered_fsl}
\title{Orders the Full Sibling List (FSL) by size and orders the indivs within each FSL for easy viewing}
\usage{
ordered_fsl(S)
}
\arguments{
\item{S}{The full sibling list}
}
\description{
This is used internally
}

