% Generated by roxygen2 (4.0.0): do not edit by hand
\name{possible_sibgroups}
\alias{possible_sibgroups}
\title{return a vector of indices of sibships that include at least one hi-sibship-lod individual}
\usage{
possible_sibgroups(IFS, AFS)
}
\arguments{
\item{IFS}{the individual full siblings group vector. A vector such that element i contains the
base-0 index of the sibship to which the individual with index i belongs.}

\item{AFS}{the vector of individual indexes (base 0) that have hi sibship lod with the
focal individual.  i.e. AFS = the Acceptable Full Siblings}
}
\description{
Note that it is currently up to the user to ensure that no element of AFS exceeds the length-1
of IFS
}
\examples{
possible_sibgroups(rep(16,50), 0:49)
possible_sibgroups(rep(0:9, each=3), sample(0:29, 10))
}

