% Generated by roxygen2 (4.0.0): do not edit by hand
\name{prep_all_variables}
\alias{prep_all_variables}
\title{a function to prep all the variables up through the marriage likelihoods}
\usage{
prep_all_variables(genos, geno_error_rates)
}
\arguments{
\item{DF}{the data frame of SNP data}

\item{geno_error_rates}{vector of genotyping error rates for each locus (recycled if necessary)}
}
\description{
a function to prep all the variables up through the marriage likelihoods
}

