\name{print.ind_geno_lik_array}
\alias{print.ind_geno_lik_array}
\title{a method for printing objects of class ind_geno_lik_array}
\usage{
\method{print}{ind_geno_lik_array}(g)
}
\arguments{
  \item{g}{An object of class ind_geno_lik_array}
}
\description{
Note that this method is called just by passing
an object of class ind_geno_lik_array to the print function.
}

