\name{print.marriage_geno_lik_array}
\alias{print.marriage_geno_lik_array}
\title{a method for printing objects of class marriage_geno_lik_array}
\usage{
\method{print}{marriage_geno_lik_array}(g)
}
\arguments{
  \item{g}{An object of class marriage_geno_lik_array}
}
\description{
Note that this method is called just by passing
an object of class marriage_geno_lik_array to the print function.
}

