% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pseudo_prior}
\alias{pseudo_prior}
\title{compute the pseudo-prior for individual and individual pulled out of sibgroup IndG}
\usage{
pseudo_prior(FSL, IndG, AFS)
}
\arguments{
\item{FSL}{the full sibling list}

\item{IndG}{The full sibling group to which the focal individual belongs}

\item{AFS}{acceptable full sibling groups for the focal individual.}
}
\value{
This returns a list of two components. $solo is the prob that the individual will form a singleton and
$afs is the prior that the individual will join any of the existing sibgroups in afs, scaled to sum to one.
}
\description{
compute the pseudo-prior for individual and individual pulled out of sibgroup IndG
}

