% Generated by roxygen2 (4.0.0): do not edit by hand
\name{run_mcmc}
\alias{run_mcmc}
\title{Run the MCMC for full sibling inference}
\usage{
run_mcmc(genos, mu = 0.005, pair_prob_cutoff = 0.001, burn_in, num_sweeps)
}
\arguments{
\item{burn_in}{Number of sweeps to discard as burn-in.  (In one sweep, every
individual in the sample has been updated once by Gibbs sampling.)}

\item{num_sweeps}{Number of sweeps after burn in to use as a sample for
MCMC.}

\item{genos}{The data frame of SNP genotypes}

\item{mu}{Genotyping error rates per locus (recycles as necessary)}

\item{pair_prob_cutoff}{the fraction of simulated full sib pairs that should have a
log-likelihood ratio for the sibling relationship less than the cutoff that will be used.
(The value pair_prob_cutoff is used to determine the which pairs have a high enough
log-likelihood ratio to be considered as possible full siblings.)}
}
\description{
Run the MCMC for full sibling inference
}
\examples{
# this is a much shorter run than one should do:
run_mcmc(fs_dev_test_data$chinook_full_sibs_genos, burn_in = 3, num_sweeps = 5)
}

