% Generated by roxygen2 (4.0.0): do not edit by hand
\name{SlurpSibships}
\alias{SlurpSibships}
\title{Slurps a colony inferred sibship file into a list}
\usage{
SlurpSibships(file, skip = integer(0))
}
\arguments{
\item{file}{the name of the colony output file.  It should be one of the ones with
the extension \code{.BestFSFamily}.}

\item{skip}{a vector of lines in the file that
should be skipped or ignored. (like maybe the first line...)}
}
\value{
A list of two named componenets:
\describe{
 \item{probs}{Colony sibship approximated posterior probababilities}
 \item{sibs}{Vectors of the individuals in each sibship}
}
}
\description{
It also  sorts it by size of sibship and
probability.
}

