% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sort_visited_sibgroups}
\alias{sort_visited_sibgroups}
\title{sorts the visited sibgroups on the basis of posterior and size}
\usage{
sort_visited_sibgroups(VS)
}
\arguments{
\item{VS}{A vector of counts of the number of times each sibship was visited.  The sibships are named
according to who is in them via the hashable_sibships function.}
}
\description{
Returns a list that has for each sibgroup a list which holds:
a vector of \code{Indivs} the number of visits, \code{Visits},
and the size of the sibgroup, \code{NumSibs}
}

