% Generated by roxygen2 (4.0.0): do not edit by hand
\name{trans_probs}
\alias{trans_probs}
\title{returns a matrix of the 27 transmission probs between a pair of parents and an offspring}
\usage{
trans_probs()
}
\description{
There are 9 possible genotype configurations among the parents and 3 in the kid, this
function just enumerates those and puts the probabilities into a 3 x 3 x 3 matrix and
returns it. It initializes all values to 0, then only modifies the non-zero ones.
}

