% Generated by roxygen2 (4.0.1): do not edit by hand
\name{update_marriage_node_kid_prongs_in_place}
\alias{update_marriage_node_kid_prongs_in_place}
\title{update marriage node kid prongs IN PLACE}
\usage{
update_marriage_node_kid_prongs_in_place(S, MP, KP, NGS_P, NGS_K, TP, bz_idx)
}
\arguments{
\item{S}{a list of lists each with two components.  The first is LMMI_Idx which is the base-0
index of the Marriage that the component is referring to, and the second is Indivs which give
the indices (base 0) of the individuals in the
full sibling groups.}

\item{MP}{the marriage posteriors matrix to be used in the udpating. Should be a matrix
that has NGS_P * L rows, where L is the number of loci.}

\item{KP}{Kid prongs.  This is what gets updated in place. Should be a matrix that has
NGS_K * L rows, where L is the number of loci.}

\item{NGS_P}{Number of genotypic states in a parent pair.  This will typically be 9.}

\item{NGS_K}{Number of genotypic states in a kid.  This will typically be 3}

\item{TP}{transmision probs as returned by trans_probs()}

\item{bz_idx}{An integer vector holding the BASE-0 indices of the components of S that
will be accessed and used to update ML.}
}
\value{
This doesn't return anything.  It modifies Prongs in place via call by reference.  Our
goal here is to make updates without copying a lot of memory.
}
\description{
This uses the information in the PMMFS matrix (the
"Posterior Matrix of Marriages given Full Siblings") to calculate
the posterior predictive distribution for the next full sibling
in the full sibling group which is referenced through S and bz_idx
}

