% Generated by roxygen2 (4.0.1): do not edit by hand
\name{update_marriage_posteriors_in_place}
\alias{update_marriage_posteriors_in_place}
\title{Calculate/update one row of a marriage Posterior matrix IN PLACE!}
\usage{
update_marriage_posteriors_in_place(ML, MP, Pri, NGS, bz_idx)
}
\arguments{
\item{ML}{the marriage likelihoods matrix}

\item{MP}{the marriage posteriors matrix that is to be updated}

\item{Pri}{the joint prior probbabilities of the parent pair}

\item{NGS}{Number of genotypic states.  For pairs of parents, for example, this will be 9.}

\item{bz_idx}{An integer vector holding the BASE-0 index of the columns of MP that needs updating.  Note that this
is not the index of the sibhip, but rather the column that sibship occupies in MP.  That will be something like
FSL[Indiv]$LMMI_Idx that}
}
\value{
This returns the value of the index of the last marriage that it updated.  It does this so I can check some things.  It modifies MP in place via call by reference.  Our
goal here is to make updates without copying a lot of memory.
}
\description{
The intended use of this is to update the marriage posteriors after an individual has
been moved from one sibship to another.  So, for example, if you moved an individual
from sibship 40 to sibship 291 (as subscripted in R) then you would want to run this
with \code{bz_idx} equal to the column of sibship 39, and then run it again with it equal
to the column of  290, after you had updated the LMMFS
for each of those.
}

