% Generated by roxygen2 (4.0.0): do not edit by hand
\name{alle_freqs}
\alias{alle_freqs}
\title{counts alleles from a geno_counts matrix}
\usage{
alle_freqs(x, proportion = T, smidge = 0.5)
}
\arguments{
  \item{x}{the geno counts matrix}

  \item{proportion}{if false it returns the counts. if true
  the proportions (summing to one)}

  \item{smidge}{small amount to add to count of each allele
  (could be useful if there are zero counts on some
  alleles)}
}
\value{
A matrix with 2 rows (corresponding to alleles 0 and 1) and L columns.
}
\description{
counts alleles from a geno_counts matrix
}

