% Generated by roxygen2 (4.0.0): do not edit by hand
\name{full_sib_mcmc_initialize}
\alias{full_sib_mcmc_initialize}
\title{initilize all the variables to do a full sibling analysis given genos and mu}
\usage{
full_sib_mcmc_initialize(genos, mu)
}
\arguments{
\item{genos}{The data frame of SNP genotypes}

\item{mu}{Genotyping error rates per locus (recycles as necessary)}
}
\description{
This returns everything in a named list suitable for do.calling
the MCMC sweep function.
}

