% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get_indiv_geno_lik}
\alias{get_indiv_geno_lik}
\title{given genotyping error and observed genotypes, return an array of likelihoods for all the individuals}
\usage{
get_indiv_geno_lik(SI, mu)
}
\arguments{
\item{SI}{a G x L x N array of snp genotype indicators, of the sort that would come out of
\code{\link{genos_to_indicators}}.  G is the number of genotypic states, L the number of loci,
and N the number of individuals.}

\item{mu}{array of per-gene-copy genotyping error rates.  Gets recycled as need be.}
}
\description{
given genotyping error and observed genotypes, return an array of likelihoods for all the individuals
}

