% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gibbs_update_one_indiv_in_place}
\alias{gibbs_update_one_indiv_in_place}
\title{does a gibbs update of the full sibling group of individual Ind}
\usage{
gibbs_update_one_indiv_in_place(FSL, IFS, LMMI, LMMFS, PMMFS, KidProngs, Pile,
  AFSL, Gfreqs, UPG, TP, IndLiks, Ind)
}
\arguments{
\item{FSL}{list of integer vectors. Each component is a list of two componontes: \code{LMMI_Idx} = the index of the sibgroup
in LMMI (see below).  It also is the column that Ind's current full sib group occuppies in LMMFS}

\item{IFS}{individual full siblings group vector. A vector such that element i contains the
base-0 index of the sibship to which the individual with index i belongs.
and PMMFS, and KidProngs \code{Indivs} = the the base-0 indices of individuals in the full sibship.}

\item{LMMI}{Likelihood matrix of marriages given individuals.}

\item{LMMFS}{Likelihood matrix of marriages given full sibships.}

\item{PMMFS}{Posterior matrix of marriages given full sibships.}

\item{KidProngs}{Posterior predictives for the next individuals to be sampled from a full sibship.}

\item{Pile}{An integer vector that we will use as a stack to hold the indices of empty marriage columns that
can be renewed.}

\item{AFSL}{The acceptable full siblings list.  Component i is a vector of the base-0 indices of the individuals
that indiv i has high full-sibling LOD with.}

\item{Gfreqs}{The genotype frequencies.}

\item{UPG}{Unrelated Pair Genotype Frequencies.}

\item{TP}{Transmission probabilities}

\item{IndLiks}{Individual likelihoods}

\item{Ind}{the index of the individual to be updated.}
}
\description{
does a gibbs update of the full sibling group of individual Ind
}

