% Generated by roxygen2 (4.0.1): do not edit by hand
\name{initialize_sib_list_to_singletons}
\alias{initialize_sib_list_to_singletons}
\title{initialize the list of full siblings to a bunch of singletons}
\usage{
initialize_sib_list_to_singletons(N)
}
\arguments{
\item{N}{number of individuals}
}
\description{
initialize the list of full siblings to a bunch of singletons
}

