% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sib_result_plot}
\alias{sib_result_plot}
\title{plot posterior bars and compare to correct or incorrect}
\usage{
sib_result_plot(SS, post, false_positive, lscale = 0.01, correct,
  add = FALSE)
}
\arguments{
\item{SS}{vector of sibsizes}

\item{post}{vector of posteriors associated with each sibship (between 0 and 1)}

\item{false_positive}{A logical vector indicating whether the sibship included
any non-sibs}

\item{lscale}{how big should the lengths be on the y-scale. i.e. one sibling = lscale}

\item{correct}{logical vector saying whether each inferred sibship is correct or not}

\item{add}{TRUE means add to existing plot}
}
\description{
Makes a nice plot.
}

